Refactor Robust Accessorial work into new file in models package.
